WORKAROUND: mm-video: venc: Correct a typo in variable name
until we have http://goo.gl/jVxRvX merged in qcom media-caf, we
can at least fix the data type to a long, but keep the name like
it is, so we dont break the build because of the missing commit,
meantioned above.

REVERT once the above commit is merged and we implement the real
patch for this.

Change-Id: I91aec18a9452feee78a00dc68d95ccd9cff22ba9